+++ CHANGELOG.md
@@ -5,7 +5,7 @@
 **New features:**
 
 * st-color-picker: Create component
-
+* st-two-list-selection: allow to show dropdown menu options  if it is a non-editable list.
 
 ## 16.4.0 (June 10, 2019)